added stance width selector
